requirements.txt
customtkinter
Pillow
pandas
openpyxl
pdf2docx
docx2pdf
pyinstaller